Use a ParameterFactory to create parameters
