+++ core/src/main/java/org/frankframework/parameters/IParameter.java
@@ -1,0 +1,5 @@
+package org.frankframework.parameters;
+
+public interface IParameter {
+
+}
