+++ core/src/main/java/org/frankframework/parameters/Parameter.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013, 2016, 2019, 2020 Nationale-Nederlanden, 2021-2023 WeAreFrank!
+   Copyright 2024 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,1034 +15,43 @@
 */
 package org.frankframework.parameters;
 
-import static org.frankframework.functional.FunctionalUtil.logValue;
-import static org.frankframework.util.StringUtil.hide;
-
-import java.io.IOException;
-import java.text.DateFormat;
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.text.MessageFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import javax.xml.transform.Source;
-import javax.xml.transform.TransformerException;
-import javax.xml.transform.dom.DOMResult;
-
-import org.apache.commons.lang3.StringUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.frankframework.configuration.ConfigurationException;
-import org.frankframework.configuration.ConfigurationUtils;
-import org.frankframework.core.IConfigurable;
-import org.frankframework.core.IWithParameters;
-import org.frankframework.core.ParameterException;
-import org.frankframework.core.PipeLineSession;
-import org.frankframework.doc.DocumentedEnum;
-import org.frankframework.doc.EnumLabel;
-import org.frankframework.jdbc.StoredProcedureQuerySender;
-import org.frankframework.pipes.PutSystemDateInSession;
-import org.frankframework.stream.Message;
-import org.frankframework.util.CredentialFactory;
-import org.frankframework.util.DateFormatUtils;
-import org.frankframework.util.DomBuilderException;
-import org.frankframework.util.EnumUtils;
-import org.frankframework.util.Misc;
-import org.frankframework.util.StringUtil;
-import org.frankframework.util.TransformerPool;
-import org.frankframework.util.TransformerPool.OutputType;
-import org.frankframework.util.UUIDUtil;
-import org.frankframework.util.XmlBuilder;
-import org.frankframework.util.XmlException;
-import org.frankframework.util.XmlUtils;
-import org.springframework.context.ApplicationContext;
-import org.w3c.dom.Document;
-import org.w3c.dom.Node;
+import org.frankframework.doc.Mandatory;
 
 import lombok.Getter;
-import lombok.Setter;
 
 /**
- * Generic parameter definition.
+ * Placeholder class to allow old-school <code>&lt;param type='number' /&gt;</code> in the new Frank!Config XSD.
  *
- * A parameter resembles an attribute. However, while attributes get their value at configuration-time,
- * parameters get their value at the time of processing the message. Value can be retrieved from the message itself,
- * a fixed value, or from the pipelineSession. If this does not result in a value (or if neither of these is specified), a default value
- * can be specified. If an XPathExpression or stylesheet is specified, it will be applied to the message, the value retrieved
- * from the pipelineSession or the fixed value specified. If the transformation produces no output, the default value
- * of the parameter is taken if provided.
- * <br/><br/>
- * Examples:
- * <pre><code>
- * stored under SessionKey 'TransportInfo':
- *  &lt;transportinfo&gt;
- *   &lt;to&gt;***@zonnet.nl&lt;/to&gt;
- *   &lt;to&gt;***@zonnet.nl&lt;/to&gt;
- *   &lt;cc&gt;***@zonnet.nl&lt;/cc&gt;
- *  &lt;/transportinfo&gt;
- *
- * to obtain all 'to' addressees as a parameter:
- * sessionKey="TransportInfo"
- * xpathExpression="transportinfo/to"
- * type="xml"
- *
- * Result:
- *   &lt;to&gt;***@zonnet.nl&lt;/to&gt;
- *   &lt;to&gt;***@zonnet.nl&lt;/to&gt;
- * </code></pre>
- *
- * N.B. to obtain a fixed value: a non-existing 'dummy' <code>sessionKey</code> in combination with the fixed value in <code>defaultValue</code> is used traditionally.
- * The current version of parameter supports the 'value' attribute, that is sufficient to set a fixed value.
- * @author Gerrit van Brakel
- * @ff.parameters Parameters themselves can have parameters too, for instance if a XSLT transformation is used, that transformation can have parameters.
+ * @author Niels Meijer
  */
-public class Parameter implements IConfigurable, IWithParameters {
-	private static final Logger LOG = LogManager.getLogger(Parameter.class);
-	private final @Getter ClassLoader configurationClassLoader = Thread.currentThread().getContextClassLoader();
-	private @Getter @Setter ApplicationContext applicationContext;
+// Should never be instantiated directly. See {@link ParameterFactory} and {@link ParameterType} for more information.
+public class Parameter extends AbstractParameter {
 
-	public static final String TYPE_DATE_PATTERN="yyyy-MM-dd";
-	public static final String TYPE_TIME_PATTERN="HH:mm:ss";
-	public static final String TYPE_DATETIME_PATTERN="yyyy-MM-dd HH:mm:ss";
-	public static final String TYPE_TIMESTAMP_PATTERN= DateFormatUtils.FORMAT_FULL_GENERIC;
-
-	public static final String FIXEDUID ="0a1b234c--56de7fa8_9012345678b_-9cd0";
-	public static final String FIXEDHOSTNAME ="MYHOST000012345";
-
-	private String name = null;
-	private @Getter ParameterType type = ParameterType.STRING;
-	private @Getter String sessionKey = null;
-	private @Getter String sessionKeyXPath = null;
-	private @Getter String contextKey = null;
-	private @Getter String xpathExpression = null;
-	private @Getter String namespaceDefs = null;
-	private @Getter String styleSheetName = null;
-	private @Getter String pattern = null;
-	private @Getter String authAlias;
-	private @Getter String username;
-	private @Getter String password;
-	private @Getter boolean ignoreUnresolvablePatternElements = false;
-	private @Getter String defaultValue = null;
-	private @Getter String defaultValueMethods = "defaultValue";
-	private @Getter String value = null;
-	private @Getter String formatString = null;
-	private @Getter String decimalSeparator = null;
-	private @Getter String groupingSeparator = null;
-	private @Getter int minLength = -1;
-	private @Getter int maxLength = -1;
-	private @Getter String minInclusiveString = null;
-	private @Getter String maxInclusiveString = null;
-	private Number minInclusive;
-	private Number maxInclusive;
-	private @Getter boolean hidden = false;
-	private @Getter boolean removeNamespaces=false;
-	private @Getter int xsltVersion = 0; // set to 0 for auto-detect.
-
-	private @Getter DecimalFormatSymbols decimalFormatSymbols = null;
-	private TransformerPool transformerPool = null;
-	private TransformerPool tpDynamicSessionKey = null;
-	protected ParameterList paramList = null;
-	private boolean configured = false;
-	private CredentialFactory cf;
-
-	private List<DefaultValueMethods> defaultValueMethodsList;
-
-	@Getter
-	private ParameterMode mode = ParameterMode.INPUT;
-
-	public enum ParameterMode {
-		INPUT, OUTPUT, INOUT
-	}
-
-	public enum DefaultValueMethods implements DocumentedEnum {
-		@EnumLabel("defaultValue")	DEFAULTVALUE,
-		@EnumLabel("sessionKey")	SESSIONKEY,
-		@EnumLabel("pattern")		PATTERN,
-		@EnumLabel("value")			VALUE,
-		@EnumLabel("input") 		INPUT;
-	}
+	private @Getter ParameterType type;
 
 	public Parameter() {
-		super();
+		//throw new NotImplementedException(); // will be replaced by appropriate executor class in ParameterFactory, based on function
 	}
 
 	/** utility constructor, useful for unit testing */
 	public Parameter(String name, String value) {
 		this();
-		this.name = name;
-		this.value = value;
-	}
-
-	@Override
-	public void addParameter(Parameter p) {
-		if (paramList==null) {
-			paramList=new ParameterList();
-		}
-		paramList.add(p);
-	}
-
-	@Override
-	public ParameterList getParameterList() {
-		return paramList;
-	}
-
-	@Override
-	public void configure() throws ConfigurationException {
-		if (StringUtils.isNotEmpty(getSessionKey()) && StringUtils.isNotEmpty(getSessionKeyXPath())) {
-			throw new ConfigurationException("Parameter ["+getName()+"] cannot have both sessionKey and sessionKeyXPath specified");
-		}
-		if (paramList!=null) {
-			paramList.configure();
-		}
-		if (StringUtils.isNotEmpty(getXpathExpression()) || StringUtils.isNotEmpty(styleSheetName)) {
-			OutputType outputType = getType() == ParameterType.XML || getType() == ParameterType.NODE || getType() == ParameterType.DOMDOC ? OutputType.XML : OutputType.TEXT;
-			boolean includeXmlDeclaration = false;
-
-			transformerPool = TransformerPool.configureTransformer0(this, getNamespaceDefs(), getXpathExpression(), getStyleSheetName(), outputType, includeXmlDeclaration, paramList, getXsltVersion());
-		} else {
-			if (paramList != null && StringUtils.isEmpty(getPattern())) {
-				throw new ConfigurationException("Parameter [" + getName() + "] can only have parameters itself if a styleSheetName, xpathExpression or pattern is specified");
-			}
-		}
-		if (StringUtils.isNotEmpty(getSessionKeyXPath())) {
-			tpDynamicSessionKey = TransformerPool.configureTransformer(this, getNamespaceDefs(), getSessionKeyXPath(), null, OutputType.TEXT,false,null);
-		}
-		if (getType() == null) {
-			LOG.info("parameter [{} has no type. Setting the type to [{}]", this::getType, ()->ParameterType.STRING);
-			setType(ParameterType.STRING);
-		}
-		if(StringUtils.isEmpty(getFormatString())) {
-			switch(getType()) {
-				case DATE:
-					setFormatString(TYPE_DATE_PATTERN);
-					break;
-				case DATETIME:
-					setFormatString(TYPE_DATETIME_PATTERN);
-					break;
-				case TIMESTAMP:
-					setFormatString(TYPE_TIMESTAMP_PATTERN);
-					break;
-				case TIME:
-					setFormatString(TYPE_TIME_PATTERN);
-					break;
-				default:
-					break;
-			}
-		}
-		if (getType()==ParameterType.NUMBER) {
-			decimalFormatSymbols = new DecimalFormatSymbols();
-			if (StringUtils.isNotEmpty(getDecimalSeparator())) {
-				decimalFormatSymbols.setDecimalSeparator(getDecimalSeparator().charAt(0));
-			}
-			if (StringUtils.isNotEmpty(getGroupingSeparator())) {
-				decimalFormatSymbols.setGroupingSeparator(getGroupingSeparator().charAt(0));
-			}
-		}
-		configured = true;
-
-		if (getMinInclusiveString()!=null || getMaxInclusiveString()!=null) {
-			if (getType()!=ParameterType.NUMBER) {
-				throw new ConfigurationException("minInclusive and maxInclusive only allowed in combination with type ["+ParameterType.NUMBER+"]");
-			}
-			if (getMinInclusiveString()!=null) {
-				DecimalFormat df = new DecimalFormat();
-				df.setDecimalFormatSymbols(decimalFormatSymbols);
-				try {
-					minInclusive = df.parse(getMinInclusiveString());
-				} catch (ParseException e) {
-					throw new ConfigurationException("Attribute [minInclusive] could not parse result ["+getMinInclusiveString()+"] to number; decimalSeparator ["+decimalFormatSymbols.getDecimalSeparator()+"] groupingSeparator ["+decimalFormatSymbols.getGroupingSeparator()+"]",e);
-				}
-			}
-			if (getMaxInclusiveString()!=null) {
-				DecimalFormat df = new DecimalFormat();
-				df.setDecimalFormatSymbols(decimalFormatSymbols);
-				try {
-					maxInclusive = df.parse(getMaxInclusiveString());
-				} catch (ParseException e) {
-					throw new ConfigurationException("Attribute [maxInclusive] could not parse result ["+getMaxInclusiveString()+"] to number; decimalSeparator ["+decimalFormatSymbols.getDecimalSeparator()+"] groupingSeparator ["+decimalFormatSymbols.getGroupingSeparator()+"]",e);
-				}
-			}
-		}
-		if (StringUtils.isNotEmpty(getAuthAlias()) || StringUtils.isNotEmpty(getUsername()) || StringUtils.isNotEmpty(getPassword())) {
-			cf=new CredentialFactory(getAuthAlias(), getUsername(), getPassword());
-		}
-	}
-
-	private List<DefaultValueMethods> getDefaultValueMethodsList() {
-		if (defaultValueMethodsList == null) {
-			defaultValueMethodsList = StringUtil.splitToStream(getDefaultValueMethods(), ", ")
-					.map(token -> EnumUtils.parse(DefaultValueMethods.class, token))
-					.collect(Collectors.toList());
-		}
-		return defaultValueMethodsList;
-	}
-
-	private Document transformToDocument(Source xmlSource, ParameterValueList pvl) throws TransformerException, IOException {
-		TransformerPool pool = getTransformerPool();
-		DOMResult transformResult = new DOMResult();
-		pool.transform(xmlSource,transformResult, pvl);
-		return (Document) transformResult.getNode();
-	}
-
-	public boolean isWildcardSessionKey() {
-		return "*".equals(getSessionKey());
-	}
-	/**
-	 * if this returns true, then the input value must be repeatable, as it might be used multiple times.
-	 */
-	public boolean requiresInputValueForResolution() {
-		if (tpDynamicSessionKey != null) { // tpDynamicSessionKey is applied to the input message to retrieve the session key
-			return true;
-		}
-		return StringUtils.isEmpty(getContextKey()) &&
-				(StringUtils.isEmpty(getSessionKey()) && StringUtils.isEmpty(getValue()) && StringUtils.isEmpty(getPattern())
-						|| getDefaultValueMethodsList().contains(DefaultValueMethods.INPUT)
-				);
-	}
-
-	public boolean requiresInputValueOrContextForResolution() {
-		if (tpDynamicSessionKey != null) { // tpDynamicSessionKey is applied to the input message to retrieve the session key
-			return true;
-		}
-		return StringUtils.isEmpty(getSessionKey()) && StringUtils.isEmpty(getValue()) && StringUtils.isEmpty(getPattern())
-					|| getDefaultValueMethodsList().contains(DefaultValueMethods.INPUT);
-	}
-
-	public boolean consumesSessionVariable(String sessionKey) {
-		return StringUtils.isEmpty(getContextKey()) && (
-					sessionKey.equals(getSessionKey())
-					|| getPattern()!=null && getPattern().contains("{"+sessionKey+"}")
-					|| getParameterList()!=null && getParameterList().consumesSessionVariable(sessionKey)
-				);
-	}
-
-	/**
-	 * determines the raw value
-	 */
-	public Object getValue(ParameterValueList alreadyResolvedParameters, Message message, PipeLineSession session, boolean namespaceAware) throws ParameterException {
-		Object result = null;
-		LOG.debug("Calculating value for Parameter [{}]", this::getName);
-		if (!configured) {
-			throw new ParameterException(getName(), "Parameter ["+getName()+"] not configured");
-		}
-
-		String requestedSessionKey;
-		if (tpDynamicSessionKey != null) {
-			try {
-				requestedSessionKey = tpDynamicSessionKey.transform(message.asSource());
-			} catch (Exception e) {
-				throw new ParameterException(getName(), "SessionKey for parameter ["+getName()+"] exception on transformation to get name", e);
-			}
-		} else {
-			requestedSessionKey = getSessionKey();
-		}
-		TransformerPool pool = getTransformerPool();
-		if (pool != null) {
-			try {
-				/*
-				 * determine source for XSLT transformation from
-				 * 1) value attribute
-				 * 2) requestedSessionKey
-				 * 3) pattern
-				 * 4) input message
-				 *
-				 * N.B. this order differs from untransformed parameters
-				 */
-				Source source;
-				if (getValue() != null) {
-					source = XmlUtils.stringToSourceForSingleUse(getValue(), namespaceAware);
-				} else if (StringUtils.isNotEmpty(requestedSessionKey)) {
-					Object sourceObject = session.get(requestedSessionKey);
-					if (getType() == ParameterType.LIST && sourceObject instanceof List) {
-						// larva can produce the sourceObject as list
-						List<String> items = (List<String>) sourceObject;
-						XmlBuilder itemsXml = new XmlBuilder("items");
-						for (String item : items) {
-							XmlBuilder itemXml = new XmlBuilder("item");
-							itemXml.setValue(item);
-							itemsXml.addSubElement(itemXml);
-						}
-						source = XmlUtils.stringToSourceForSingleUse(itemsXml.toXML(), namespaceAware);
-					} else if (getType() == ParameterType.MAP && sourceObject instanceof Map) {
-						// larva can produce the sourceObject as map
-						Map<String, String> items = (Map<String, String>) sourceObject;
-						XmlBuilder itemsXml = new XmlBuilder("items");
-						for (String item : items.keySet()) {
-							XmlBuilder itemXml = new XmlBuilder("item");
-							itemXml.addAttribute("name", item);
-							itemXml.setValue(items.get(item));
-							itemsXml.addSubElement(itemXml);
-						}
-						source = XmlUtils.stringToSourceForSingleUse(itemsXml.toXML(), namespaceAware);
-					} else {
-						Message sourceMsg = Message.asMessage(sourceObject);
-						if (StringUtils.isNotEmpty(getContextKey())) {
-							sourceMsg = Message.asMessage(sourceMsg.getContext().get(getContextKey()));
-						}
-						if (!sourceMsg.isEmpty()) {
-							LOG.debug("Parameter [{}] using sessionvariable [{}] as source for transformation", this::getName, () -> requestedSessionKey);
-							source = sourceMsg.asSource();
-						} else {
-							LOG.debug("Parameter [{}] sessionvariable [{}] empty, no transformation will be performed", this::getName, () -> requestedSessionKey);
-							source = null;
-						}
-					}
-				} else if (StringUtils.isNotEmpty(getPattern())) {
-					String sourceString = formatPattern(alreadyResolvedParameters, session);
-					if (StringUtils.isNotEmpty(sourceString)) {
-						LOG.debug("Parameter [{}] using pattern [{}] as source for transformation", this::getName, this::getPattern);
-						source = XmlUtils.stringToSourceForSingleUse(sourceString, namespaceAware);
-					} else {
-						LOG.debug("Parameter [{}] pattern [{}] empty, no transformation will be performed", this::getName, this::getPattern);
-						source = null;
-					}
-				} else if (message != null) {
-					if (StringUtils.isNotEmpty(getContextKey())) {
-						source = Message.asMessage(message.getContext().get(getContextKey())).asSource();
-					} else {
-						source = message.asSource();
-					}
-				} else {
-					source = null;
-				}
-				if (source!=null) {
-					if (isRemoveNamespaces()) {
-						// TODO: There should be a more efficient way to do this
-						String rnResult = XmlUtils.removeNamespaces(XmlUtils.source2String(source));
-						source = XmlUtils.stringToSource(rnResult);
-					}
-					ParameterValueList pvl = paramList == null ? null : paramList.getValues(message, session, namespaceAware);
-					switch (getType()) {
-						case NODE:
-							return transformToDocument(source, pvl).getFirstChild();
-						case DOMDOC:
-							return transformToDocument(source, pvl);
-						default:
-							String transformResult = pool.transform(source, pvl);
-							if (StringUtils.isNotEmpty(transformResult)) {
-								result = transformResult;
-							}
-							break;
-					}
-				}
-			} catch (Exception e) {
-				throw new ParameterException(getName(), "Parameter ["+getName()+"] exception on transformation to get parametervalue", e);
-			}
-		} else {
-			/*
-			 * No XSLT transformation, determine primary result from
-			 * 1) requestedSessionKey
-			 * 2) pattern
-			 * 3) value attribute
-			 * 4) input message
-			 *
-			 * N.B. this order differs from transformed parameters.
-			 */
-			if (StringUtils.isNotEmpty(requestedSessionKey)) {
-				result = session.get(requestedSessionKey);
-				if (result instanceof Message message1 && StringUtils.isNotEmpty(getContextKey())) {
-					result = message1.getContext().get(getContextKey());
-				}
-				if (LOG.isDebugEnabled() && (result == null ||
-						((result instanceof String string) && string.isEmpty()) ||
-						((result instanceof Message message1) && message1.isEmpty()))) {
-					LOG.debug("Parameter [{}] session variable [{}] is empty", this::getName, () -> requestedSessionKey);
-				}
-			} else if (StringUtils.isNotEmpty(getPattern())) {
-				result = formatPattern(alreadyResolvedParameters, session);
-			} else if (getValue()!=null) {
-				result = getValue();
-			} else {
-				try {
-					if (message==null) {
-						return null;
-					}
-					if (StringUtils.isNotEmpty(getContextKey())) {
-						result = message.getContext().get(getContextKey());
-					} else {
-						message.preserve();
-						result=message;
-					}
-				} catch (IOException e) {
-					throw new ParameterException(getName(), e);
-				}
-			}
-		}
-
-		if (result instanceof Message resultMessage) {
-			if (Message.isNull(resultMessage)) {
-				result = null;
-			} else if (resultMessage.isRequestOfType(String.class)) { //Used by getMinLength and getMaxLength
-				try {
-					result = resultMessage.asString();
-				} catch (IOException ignored) {
-					// Already checked for String, so this should never happen
-				}
-			}
-		}
-		if (result != null && !"".equals(result)) {
-			final Object finalResult = result;
-			LOG.debug("Parameter [{}] resolved to [{}]", this::getName, ()-> isHidden() ? hide(finalResult.toString()) : finalResult);
-		} else {
-			// if result is empty then return specified default value
-			Object valueByDefault=null;
-			Iterator<DefaultValueMethods> it = getDefaultValueMethodsList().iterator();
-			while (valueByDefault == null && it.hasNext()) {
-				DefaultValueMethods method = it.next();
-				switch(method) {
-					case DEFAULTVALUE:
-						valueByDefault = getDefaultValue();
-						break;
-					case SESSIONKEY:
-						valueByDefault = session.get(requestedSessionKey);
-						break;
-					case PATTERN:
-						valueByDefault = formatPattern(alreadyResolvedParameters, session);
-						break;
-					case VALUE:
-						valueByDefault = getValue();
-						break;
-					case INPUT:
-						try {
-							message.preserve();
-							valueByDefault=message.asString();
-						} catch (IOException e) {
-							throw new ParameterException(getName(), e);
-						}
-						break;
-					default:
-						throw new IllegalArgumentException("Unknown defaultValues method ["+method+"]");
-				}
-			}
-			if (valueByDefault!=null) {
-				result = valueByDefault;
-				final Object finalResult = result;
-				LOG.debug("Parameter [{}] resolved to default value [{}]", this::getName, ()-> isHidden() ? hide(finalResult.toString()) : finalResult);
-			}
-		}
-		if (result instanceof String stringResult) {
-			if (getMinLength()>=0 && getType()!=ParameterType.NUMBER) {
-				if (stringResult.length() < getMinLength()) {
-					LOG.debug("Padding parameter [{}] because length [{}] falls short of minLength [{}]", this::getName, stringResult::length, this::getMinLength);
-					result = StringUtils.rightPad(stringResult, getMinLength());
-				}
-			}
-			if (getMaxLength()>=0) {
-				if (stringResult.length() > getMaxLength()) {
-					LOG.debug("Trimming parameter [{}] because length [{}] exceeds maxLength [{}]", this::getName, stringResult::length, this::getMaxLength);
-					result = stringResult.substring(0, getMaxLength());
-				}
-			}
-		}
-		if(result !=null && getType().requiresTypeConversion) {
-			result = getValueAsType(result, namespaceAware);
-		}
-		if (result instanceof Number number) {
-			if (getMinInclusiveString()!=null && number.floatValue() < minInclusive.floatValue()) {
-				LOG.debug("Replacing parameter [{}] because value [{}] falls short of minInclusive [{}]", this::getName, logValue(result), this::getMinInclusiveString);
-				result = minInclusive;
-			}
-			if (getMaxInclusiveString()!=null && number.floatValue() > maxInclusive.floatValue()) {
-				LOG.debug("Replacing parameter [{}] because value [{}] exceeds maxInclusive [{}]", this::getName, logValue(result), this::getMaxInclusiveString);
-				result = maxInclusive;
-			}
-		}
-		if (getType()==ParameterType.NUMBER && getMinLength()>=0 && (result+"").length()<getMinLength()) {
-			LOG.debug("Adding leading zeros to parameter [{}]", this::getName);
-			result = StringUtils.leftPad(result+"", getMinLength(), '0');
-		}
-		return result;
-	}
-
-	/** Converts raw data to configured parameter type */
-	private Object getValueAsType(Object request, boolean namespaceAware) throws ParameterException {
-		Message requestMessage = Message.asMessage(request);
-		Object result = request;
-		try {
-			switch(getType()) {
-				case NODE:
-					try {
-						if (isRemoveNamespaces()) {
-							requestMessage = XmlUtils.removeNamespaces(requestMessage);
-						}
-						if(request instanceof Document document) {
-							return document.getDocumentElement();
-						}
-						if(request instanceof Node) {
-							return request;
-						}
-						result = XmlUtils.buildDomDocument(requestMessage.asInputSource(), namespaceAware).getDocumentElement();
-						final Object finalResult = result;
-						LOG.debug("final result [{}][{}]", ()->finalResult.getClass().getName(), ()-> finalResult);
-					} catch (DomBuilderException | IOException | XmlException e) {
-						throw new ParameterException(getName(), "Parameter ["+getName()+"] could not parse result ["+requestMessage+"] to XML nodeset",e);
-					}
-					break;
-				case DOMDOC:
-					try {
-						if (isRemoveNamespaces()) {
-							requestMessage = XmlUtils.removeNamespaces(requestMessage);
-						}
-						if(request instanceof Document) {
-							return request;
-						}
-						result = XmlUtils.buildDomDocument(requestMessage.asInputSource(), namespaceAware);
-						final Object finalResult = result;
-						LOG.debug("final result [{}][{}]", ()->finalResult.getClass().getName(), ()-> finalResult);
-					} catch (DomBuilderException | IOException | XmlException e) {
-						throw new ParameterException(getName(), "Parameter ["+getName()+"] could not parse result ["+requestMessage+"] to XML document",e);
-					}
-					break;
-				case DATE:
-				case DATETIME:
-				case TIMESTAMP:
-				case TIME: {
-					if (request instanceof Date) {
-						return request;
-					}
-					Message finalRequestMessage = requestMessage;
-					LOG.debug("Parameter [{}] converting result [{}] to Date using formatString [{}]", this::getName, () -> finalRequestMessage, this::getFormatString);
-					DateFormat df = new SimpleDateFormat(getFormatString());
-					try {
-						result = df.parseObject(requestMessage.asString());
-					} catch (ParseException e) {
-						throw new ParameterException(getName(), "Parameter [" + getName() + "] could not parse result [" + requestMessage + "] to Date using formatString [" + getFormatString() + "]", e);
-					}
-					break;
-				}
-				case XMLDATETIME: {
-					if (request instanceof Date) {
-						return request;
-					}
-					Message finalRequestMessage = requestMessage;
-					LOG.debug("Parameter [{}] converting result [{}] from XML dateTime to Date", this::getName, () -> finalRequestMessage);
-					result = XmlUtils.parseXmlDateTime(requestMessage.asString());
-					break;
-				}
-				case NUMBER: {
-					if (request instanceof Number) {
-						return request;
-					}
-					Message finalRequestMessage = requestMessage;
-					LOG.debug("Parameter [{}] converting result [{}] to number decimalSeparator [{}] groupingSeparator [{}]", this::getName, ()->finalRequestMessage, decimalFormatSymbols::getDecimalSeparator, decimalFormatSymbols::getGroupingSeparator);
-					DecimalFormat decimalFormat = new DecimalFormat();
-					decimalFormat.setDecimalFormatSymbols(decimalFormatSymbols);
-					try {
-						result = decimalFormat.parse(requestMessage.asString());
-					} catch (ParseException e) {
-						throw new ParameterException(getName(), "Parameter [" + getName() + "] could not parse result [" + requestMessage + "] to number decimalSeparator [" + decimalFormatSymbols.getDecimalSeparator() + "] groupingSeparator [" + decimalFormatSymbols.getGroupingSeparator() + "]", e);
-					}
-					break;
-				}
-				case INTEGER: {
-					if (request instanceof Integer) {
-						return request;
-					}
-					Message finalRequestMessage = requestMessage;
-					LOG.debug("Parameter [{}] converting result [{}] to integer", this::getName, ()->finalRequestMessage);
-					try {
-						result = Integer.parseInt(requestMessage.asString());
-					} catch (NumberFormatException e) {
-						throw new ParameterException(getName(), "Parameter [" + getName() + "] could not parse result [" + requestMessage + "] to integer", e);
-					}
-					break;
-				}
-				case BOOLEAN: {
-					if (request instanceof Boolean) {
-						return request;
-					}
-					Message finalRequestMessage = requestMessage;
-					LOG.debug("Parameter [{}] converting result [{}] to boolean", this::getName, ()->finalRequestMessage);
-					result = Boolean.parseBoolean(requestMessage.asString());
-					break;
-				}
-				default:
-					break;
-			}
-		} catch(IOException e) {
-			throw new ParameterException(getName(), "Could not convert parameter ["+getName()+"] to String", e);
-		}
-
-		return result;
-	}
-
-	private String formatPattern(ParameterValueList alreadyResolvedParameters, PipeLineSession session) throws ParameterException {
-		int startNdx;
-		int endNdx = 0;
-
-		// replace the named parameter with numbered parameters
-		StringBuilder formatPattern = new StringBuilder();
-		List<Object> params = new ArrayList<>();
-		int paramPosition = 0;
-		while(true) {
-			// get name of parameter in pattern to be substituted
-			startNdx = pattern.indexOf("{", endNdx);
-			if (startNdx == -1) {
-				formatPattern.append(pattern.substring(endNdx));
-				break;
-			}
-			else {
-				formatPattern.append(pattern, endNdx, startNdx);
-			}
-			int tmpEndNdx = pattern.indexOf("}", startNdx);
-			endNdx = pattern.indexOf(",", startNdx);
-			if (endNdx == -1 || endNdx > tmpEndNdx) {
-				endNdx = tmpEndNdx;
-			}
-			if (endNdx == -1) {
-				throw new ParameterException(getName(), new ParseException("Bracket is not closed", startNdx));
-			}
-			String substitutionPattern = pattern.substring(startNdx + 1, tmpEndNdx);
-
-			// get value
-			Object substitutionValue = getValueForFormatting(alreadyResolvedParameters, session, substitutionPattern);
-			params.add(substitutionValue);
-			formatPattern.append('{').append(paramPosition++);
-		}
-		try {
-			return MessageFormat.format(formatPattern.toString(), params.toArray());
-		} catch (Exception e) {
-			throw new ParameterException(getName(), "Cannot parse ["+formatPattern.toString()+"]", e);
-		}
-	}
-
-	private Object preFormatDateType(Object rawValue, String formatType, String patternFormatString) throws ParameterException {
-		if (formatType!=null && ("date".equalsIgnoreCase(formatType) || "time".equalsIgnoreCase(formatType))) {
-			if (rawValue instanceof Date) {
-				return rawValue;
-			}
-			DateFormat df = new SimpleDateFormat(StringUtils.isNotEmpty(patternFormatString) ? patternFormatString : DateFormatUtils.FORMAT_DATETIME_GENERIC);
-			try {
-				return df.parse(Message.asString(rawValue));
-			} catch (ParseException | IOException e) {
-				throw new ParameterException(getName(), "Cannot parse ["+rawValue+"] as date", e);
-			}
-		}
-		if (rawValue instanceof Date) {
-			DateFormat df = new SimpleDateFormat(StringUtils.isNotEmpty(patternFormatString) ? patternFormatString : DateFormatUtils.FORMAT_DATETIME_GENERIC);
-			return df.format(rawValue);
-		}
-		try {
-			return Message.asString(rawValue);
-		} catch (IOException e) {
-			throw new ParameterException(getName(), "Cannot read date value ["+rawValue+"]", e);
-		}
-	}
-
-
-	private Object getValueForFormatting(ParameterValueList alreadyResolvedParameters, PipeLineSession session, String targetPattern) throws ParameterException {
-		String[] patternElements = targetPattern.split(",");
-		String name = patternElements[0].trim();
-		String formatType = patternElements.length>1 ? patternElements[1].trim() : null;
-		String formatString = patternElements.length>2 ? patternElements[2].trim() : null;
-
-		ParameterValue paramValue = alreadyResolvedParameters.get(name);
-		Object substitutionValue = paramValue == null ? null : paramValue.getValue();
-
-		if (substitutionValue == null) {
-			Object substitutionValueMessage = session.get(name);
-			if (substitutionValueMessage != null) {
-				if (substitutionValueMessage instanceof Date) {
-					substitutionValue = preFormatDateType(substitutionValueMessage, formatType, formatString);
-				} else {
-					if (substitutionValueMessage instanceof String) {
-						substitutionValue = substitutionValueMessage;
-					} else {
-						try {
-							Message message = Message.asMessage(substitutionValueMessage); // Do not close object from session here; might be reused later
-							substitutionValue = message.asString();
-						} catch (IOException e) {
-							throw new ParameterException(getName(), "Cannot get substitution value from session key: " + name, e);
-						}
-					}
-					if (substitutionValue == null) throw new ParameterException(getName(), "Cannot get substitution value from session key: " + name);
-				}
-			}
-		}
-		if (substitutionValue == null) {
-			String namelc=name.toLowerCase();
-			switch (namelc) {
-				case "now":
-					substitutionValue = preFormatDateType(new Date(), formatType, formatString);
-					break;
-				case "uid":
-					substitutionValue = UUIDUtil.createSimpleUUID();
-					break;
-				case "uuid":
-					substitutionValue = UUIDUtil.createRandomUUID();
-					break;
-				case "hostname":
-					substitutionValue = Misc.getHostname();
-					break;
-				case "fixeddate":
-					if (!ConfigurationUtils.isConfigurationStubbed(configurationClassLoader)) {
-						throw new ParameterException(getName(), "Parameter pattern [" + name + "] only allowed in stub mode");
-					}
-					Object fixedDateTime = session.get(PutSystemDateInSession.FIXEDDATE_STUB4TESTTOOL_KEY);
-					if (fixedDateTime == null) {
-						DateFormat df = new SimpleDateFormat(DateFormatUtils.FORMAT_DATETIME_GENERIC);
-						try {
-							fixedDateTime = df.parse(PutSystemDateInSession.FIXEDDATETIME);
-						} catch (ParseException e) {
-							throw new ParameterException(getName(), "Could not parse FIXEDDATETIME [" + PutSystemDateInSession.FIXEDDATETIME + "]", e);
-						}
-					}
-					substitutionValue = preFormatDateType(fixedDateTime, formatType, formatString);
-					break;
-				case "fixeduid":
-					if (!ConfigurationUtils.isConfigurationStubbed(configurationClassLoader)) {
-						throw new ParameterException(getName(), "Parameter pattern [" + name + "] only allowed in stub mode");
-					}
-					substitutionValue = FIXEDUID;
-					break;
-				case "fixedhostname":
-					if (!ConfigurationUtils.isConfigurationStubbed(configurationClassLoader)) {
-						throw new ParameterException(getName(), "Parameter pattern [" + name + "] only allowed in stub mode");
-					}
-					substitutionValue = FIXEDHOSTNAME;
-					break;
-				case "username":
-					substitutionValue = cf != null ? cf.getUsername() : "";
-					break;
-				case "password":
-					substitutionValue = cf != null ? cf.getPassword() : "";
-					break;
-			}
-		}
-		if (substitutionValue == null) {
-			if (isIgnoreUnresolvablePatternElements()) {
-				substitutionValue="";
-			} else {
-				throw new ParameterException(getName(), "Parameter or session variable with name [" + name + "] in pattern [" + getPattern() + "] cannot be resolved");
-			}
-		}
-		return substitutionValue;
+		setName(name);
+		setValue(value);
 	}
 
 	@Override
 	public String toString() {
-		return "Parameter name=[" + name + "] defaultValue=[" + defaultValue + "] sessionKey=[" + sessionKey + "] sessionKeyXPath=[" + sessionKeyXPath + "] xpathExpression=[" + xpathExpression + "] type=[" + type + "] value=[" + value + "]";
+		StringBuilder builder = new StringBuilder(super.toString());
+		if(type != null) builder.append(" type ["+type+"]");
+		return builder.toString();
 	}
 
-	private TransformerPool getTransformerPool() {
-		return transformerPool;
-	}
-
-	/** Name of the parameter */
 	@Override
-	public void setName(String parameterName) {
-		name = parameterName;
-	}
-	@Override
-	public String getName() {
-		return name;
-	}
-
+	@Mandatory
 	/** The target data type of the parameter, related to the database or XSLT stylesheet to which the parameter is applied. */
 	public void setType(ParameterType type) {
 		this.type = type;
-	}
-
-	/** The value of the parameter, or the base for transformation using xpathExpression or stylesheet, or formatting. */
-	public void setValue(String value) {
-		this.value = value;
-	}
-
-	/**
-	 * Key of a PipelineSession-variable. <br/>If specified, the value of the PipelineSession variable is used as input for
-	 * the xpathExpression or stylesheet, instead of the current input message. <br/>If no xpathExpression or stylesheet are
-	 * specified, the value itself is returned. <br/>If the value '*' is specified, all existing sessionkeys are added as
-	 * parameter of which the name starts with the name of this parameter. <br/>If also the name of the parameter has the
-	 * value '*' then all existing sessionkeys are added as parameter (except tsReceived)
-	 */
-	public void setSessionKey(String string) {
-		sessionKey = string;
-	}
-
-	/** key of message context variable to use as source, instead of the message found from input message or sessionKey itself */
-	public void setContextKey(String string) {
-		contextKey = string;
-	}
-
-	/** Instead of a fixed <code>sessionKey</code> it's also possible to use a XPath expression applied to the input message to extract the name of the session-variable. */
-	public void setSessionKeyXPath(String string) {
-		sessionKeyXPath = string;
-	}
-
-	/** URL to a stylesheet that wil be applied to the contents of the message or the value of the session-variable. */
-	public void setStyleSheetName(String stylesheetName){
-		this.styleSheetName=stylesheetName;
-	}
-
-	/** the XPath expression to extract the parameter value from the (xml formatted) input or session-variable. */
-	public void setXpathExpression(String xpathExpression) {
-		this.xpathExpression = xpathExpression;
-	}
-
-	/**
-	 * If set to <code>2</code> or <code>3</code> a Saxon (net.sf.saxon) xslt processor 2.0 or 3.0 respectively will be used, otherwise xslt processor 1.0 (org.apache.xalan). <code>0</code> will auto-detect
-	 * @ff.default 0
-	 */
-	public void setXsltVersion(int xsltVersion) {
-		this.xsltVersion=xsltVersion;
-	}
-
-	/**
-	 * Namespace definitions for xpathExpression. Must be in the form of a comma or space separated list of
-	 * <code>prefix=namespaceuri</code>-definitions. One entry can be without a prefix, that will define the default namespace.
-	 */
-	public void setNamespaceDefs(String namespaceDefs) {
-		this.namespaceDefs = namespaceDefs;
-	}
-
-	/**
-	 * When set <code>true</code> namespaces (and prefixes) in the input message are removed before the stylesheet/xpathExpression is executed
-	 * @ff.default <code>false</code>
-	 */
-	public void setRemoveNamespaces(boolean b) {
-		removeNamespaces = b;
-	}
-
-	/** If the result of sessionKey, xpathExpression and/or stylesheet returns null or an empty string, this value is returned */
-	public void setDefaultValue(String string) {
-		defaultValue = string;
-	}
-
-	/**
-	 * Comma separated list of methods (<code>defaultValue</code>, <code>sessionKey</code>, <code>pattern</code>, <code>value</code> or <code>input</code>) to use as default value. Used in the order they appear until a non-null value is found.
-	 * @ff.default <code>defaultValue</code>
-	 */
-	public void setDefaultValueMethods(String string) {
-		defaultValueMethods = string;
-	}
-
-	/**
-	 * Value of parameter is determined using substitution and formatting, following MessageFormat syntax with named parameters. The expression can contain references
-	 * to <code>session-variables</code> or other <code>parameters</code> using the {name-of-parameter} and is formatted using java.text.MessageFormat.
-	 * <br/><b>NB: When referencing other <code>parameters</code> these MUST be defined before the parameter using pattern substitution.</b>
-	 * <br/>
-	 * <br/>
-	 * If for instance <code>fname</code> is a parameter or session-variable that resolves to Eric, then the pattern
-	 * 'Hi {fname}, how do you do?' resolves to 'Hi Eric, do you do?'.<br/>
-	 * The following predefined reference can be used in the expression too:<ul>
-	 * <li>{now}: the current system time</li>
-	 * <li>{uid}: an unique identifier, based on the IP address and java.rmi.server.UID</li>
-	 * <li>{uuid}: an unique identifier, based on the IP address and java.util.UUID</li>
-	 * <li>{hostname}: the name of the machine the application runs on</li>
-	 * <li>{username}: username from the credentials found using authAlias, or the username attribute</li>
-	 * <li>{password}: password from the credentials found using authAlias, or the password attribute</li>
-	 * <li>{fixeddate}: fake date, for testing only</li>
-	 * <li>{fixeduid}: fake uid, for testing only</li>
-	 * <li>{fixedhostname}: fake hostname, for testing only</li>
-	 * </ul>
-	 * A guid can be generated using {hostname}_{uid}, see also
-	 * <a href="http://java.sun.com/j2se/1.4.2/docs/api/java/rmi/server/uid.html">http://java.sun.com/j2se/1.4.2/docs/api/java/rmi/server/uid.html</a> for more information about (g)uid's or
-	 * <a href="http://docs.oracle.com/javase/1.5.0/docs/api/java/util/uuid.html">http://docs.oracle.com/javase/1.5.0/docs/api/java/util/uuid.html</a> for more information about uuid's.
-	 * <br/>
-	 * When combining a date or time <code>pattern</code> like {now} or {fixeddate} with a DATE, TIME, DATETIME or TIMESTAMP <code>type</code>, the effective value of the attribute
-	 * <code>formatString</code> must match the effective value of the formatString in the <code>pattern</code>.
-	 */
-	public void setPattern(String string) {
-		pattern = string;
-	}
-
-	/** Alias used to obtain username and password, used when a <code>pattern</code> containing {username} or {password} is specified */
-	public void setAuthAlias(String string) {
-		authAlias = string;
-	}
-
-	/** Default username that is used when a <code>pattern</code> containing {username} is specified */
-	public void setUsername(String string) {
-		username = string;
-	}
-
-	/** Default password that is used when a <code>pattern</code> containing {password} is specified */
-	public void setPassword(String string) {
-		password = string;
-	}
-
-	/** If set <code>true</code> pattern elements that cannot be resolved to a parameter or sessionKey are silently resolved to an empty string */
-	public void setIgnoreUnresolvablePatternElements(boolean b) {
-		ignoreUnresolvablePatternElements = b;
-	}
-
-	/**
-	 * Used in combination with types <code>DATE</code>, <code>TIME</code>, <code>DATETIME</code> and <code>TIMESTAMP</code> to parse the raw parameter string data into an object of the respective type
-	 * @ff.default depends on type
-	 */
-	public void setFormatString(String string) {
-		formatString = string;
-	}
-
-	/**
-	 * Used in combination with type <code>NUMBER</code>
-	 * @ff.default system default
-	 */
-	public void setDecimalSeparator(String string) {
-		decimalSeparator = string;
-	}
-
-	/**
-	 * Used in combination with type <code>NUMBER</code>
-	 * @ff.default system default
-	 */
-	public void setGroupingSeparator(String string) {
-		groupingSeparator = string;
-	}
-
-	/**
-	 * If set (>=0) and the length of the value of the parameter falls short of this minimum length, the value is padded
-	 * @ff.default -1
-	 */
-	public void setMinLength(int i) {
-		minLength = i;
-	}
-
-	/**
-	 * If set (>=0) and the length of the value of the parameter exceeds this maximum length, the length is trimmed to this maximum length
-	 * @ff.default -1
-	 */
-	public void setMaxLength(int i) {
-		maxLength = i;
-	}
-
-	/** Used in combination with type <code>number</code>; if set and the value of the parameter exceeds this maximum value, this maximum value is taken */
-	public void setMaxInclusive(String string) {
-		maxInclusiveString = string;
-	}
-
-	/** Used in combination with type <code>number</code>; if set and the value of the parameter falls short of this minimum value, this minimum value is taken */
-	public void setMinInclusive(String string) {
-		minInclusiveString = string;
-	}
-
-	/**
-	 * If set to <code>true</code>, the value of the parameter will not be shown in the log (replaced by asterisks)
-	 * @ff.default <code>false</code>
-	 */
-	public void setHidden(boolean b) {
-		hidden = b;
-	}
-
-	/**
-	 * Set the mode of the parameter, which determines if the parameter is an INPUT, OUTPUT, or INOUT.
-	 * This parameter only has effect for {@link StoredProcedureQuerySender}.
-	 * An OUTPUT parameter does not need to have a value specified, but does need to have the type specified.
-	 * Parameter values will not be updated, but output values will be put into the result of the
-	 * {@link StoredProcedureQuerySender}.
-	 * <b/>
-	 * If not specified, the default is INPUT.
-	 *
-	 * @param mode INPUT, OUTPUT or INOUT.
-	 */
-	public void setMode(ParameterMode mode) {
-		this.mode = mode;
+		super.setType(type);
 	}
 }
